+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+- Fix: Add MongoDB command details for update and delete operations.
+
 ## 1.71.2
 - Avoid triggering fastify's `basePath` deprecation warning.